--- src/main/java/Command.java
@@ -1,9 +1,0 @@
-public abstract class Command {
-
-    abstract void execute(TaskList tasks, UI ui, Storage storage) throws TonyException;
-
-    boolean isExit() {
-        return false;
-    }
-
-}
