--- src/main/java/DeadlineException.java
@@ -1,6 +1,0 @@
-public class DeadlineException extends TonyException {
-
-    public DeadlineException(String string) {
-        super(string);
-    }
-}
