--- src/main/java/DeleteException.java
@@ -1,6 +1,0 @@
-public class DeleteException extends TonyException {
-
-    public DeleteException(String string) {
-        super(string);
-    }
-}
