--- src/main/java/EmptyTaskException.java
@@ -1,6 +1,0 @@
-public class EmptyTaskException extends TonyException {
-
-    public EmptyTaskException(String string) {
-        super(string);
-    }
-}
