--- src/main/java/EventException.java
@@ -1,6 +1,0 @@
-public class EventException extends TonyException {
-
-    public EventException(String string) {
-        super(string);
-    }
-}
