--- src/main/java/ExitCommand.java
@@ -1,11 +1,0 @@
-public class ExitCommand extends Command {
-    @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
-
-    }
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
