--- src/main/java/InvalidCommandException.java
@@ -1,6 +1,0 @@
-public class InvalidCommandException extends TonyException {
-
-    public InvalidCommandException(String string) {
-        super(string);
-    }
-}
