--- src/main/java/InvalidIndexException.java
@@ -1,6 +1,0 @@
-public class InvalidIndexException extends TonyException {
-
-    public InvalidIndexException(String string) {
-        super(string);
-    }
-}
