--- src/main/java/MarkException.java
@@ -1,6 +1,0 @@
-public class MarkException extends TonyException {
-
-    public MarkException(String string) {
-        super(string);
-    }
-}
