--- src/main/java/ShowTasksException.java
@@ -1,6 +1,0 @@
-public class ShowTasksException extends TonyException {
-
-    public ShowTasksException(String string) {
-        super(string);
-    }
-}
