--- src/main/java/Storage.java
@@ -1,83 +1,0 @@
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.LocalDateTime;
-
-public class Storage {
-
-    private final File saveFile;
-
-    public Storage(String saveFilePath) {
-        this.saveFile = new File(saveFilePath);
-        createIfNotExists();
-    }
-
-    private void createIfNotExists() {
-        try {
-            File parentDir = saveFile.getParentFile();
-            if (parentDir != null && !parentDir.exists()) {
-                parentDir.mkdirs();
-            }
-            if (!saveFile.exists()) {
-                saveFile.createNewFile();
-            }
-        } catch (IOException e) {
-            System.out.println("Error creating save file: " + e.getMessage());
-        }
-    }
-
-    public void save(TaskList list) {
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter(saveFile))) {
-            for (Task task : list.getList()) {
-                writer.write(task.toDataFormat());
-                writer.newLine();
-            }
-        } catch (IOException e) {
-            System.out.println("Error writing to save file: " + e.getMessage());
-        }
-    }
-
-    public void load(TaskList list) {
-        try (BufferedReader reader = new BufferedReader(new FileReader(saveFile))) {
-            String line;
-            while ((line = reader.readLine()) != null) {
-                line = line.trim();
-                if (line.isEmpty()) continue;
-
-                String[] parts = line.split(" \\| ");
-                String type = parts[0];
-                boolean isDone = parts[1].equals("1");
-                Task task = null;
-
-                switch (type) {
-                    case "T":
-                        task = new ToDo(parts[2]);
-                        break;
-
-                    case "D":
-                        // parts[3] contains the deadline string
-                        task = new Deadline(parts[2], DateTimeManager.parse(parts[3]));
-                        break;
-
-                    case "E":
-                        // parts[3] contains the event duration: "from-to"
-                        String[] duration = parts[3].split(" - ", 2); // split only once
-                        String fromStr = duration[0].replace("from ", "").trim();
-                        String toStr = duration[1].replace("to ", "").trim();
-                        task = new Event(parts[2], DateTimeManager.parse(fromStr), DateTimeManager.parse(toStr));
-                        break;
-                }
-                if (task != null) {
-                    if (isDone) task.markDone();
-                    list.addTask(task);
-                }
-            }
-        } catch (IOException e) {
-            System.out.println("Error loading save file: " + e.getMessage());
-        }
-    }
-}
-
