--- src/main/java/ToDo.java
@@ -1,14 +1,0 @@
-public class ToDo extends Task {
-
-    public ToDo(String command) {
-        super(command);
-    }
-
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    public String toDataFormat() {
-        return "T | " + (doneStatus() ? 1 : 0) + " | " + this.command;
-    }
-}
