--- src/main/java/TonyException.java
@@ -1,7 +1,0 @@
-public class TonyException extends Exception {
-    private static final String border = "\t" + "+".repeat(80);
-
-    public TonyException(String string) {
-        super(border + "\n\t" + string + "\n" + border);
-    }
-}
