+++ src/main/java/tony/exceptions/DeadlineException.java
@@ -1,0 +1,8 @@
+package tony.exceptions;
+
+public class DeadlineException extends TonyException {
+
+    public DeadlineException(String string) {
+        super(string);
+    }
+}
