+++ src/main/java/tony/exceptions/DeleteException.java
@@ -1,0 +1,8 @@
+package tony.exceptions;
+
+public class DeleteException extends TonyException {
+
+    public DeleteException(String string) {
+        super(string);
+    }
+}
