+++ src/main/java/tony/exceptions/EmptyTaskException.java
@@ -1,0 +1,8 @@
+package tony.exceptions;
+
+public class EmptyTaskException extends TonyException {
+
+    public EmptyTaskException(String string) {
+        super(string);
+    }
+}
