+++ src/main/java/tony/exceptions/EventException.java
@@ -1,0 +1,8 @@
+package tony.exceptions;
+
+public class EventException extends TonyException {
+
+    public EventException(String string) {
+        super(string);
+    }
+}
