+++ src/main/java/tony/exceptions/InvalidCommandException.java
@@ -1,0 +1,8 @@
+package tony.exceptions;
+
+public class InvalidCommandException extends TonyException {
+
+    public InvalidCommandException(String string) {
+        super(string);
+    }
+}
