+++ src/main/java/tony/exceptions/InvalidIndexException.java
@@ -1,0 +1,8 @@
+package tony.exceptions;
+
+public class InvalidIndexException extends TonyException {
+
+    public InvalidIndexException(String string) {
+        super(string);
+    }
+}
