+++ src/main/java/tony/exceptions/MarkException.java
@@ -1,0 +1,8 @@
+package tony.exceptions;
+
+public class MarkException extends TonyException {
+
+    public MarkException(String string) {
+        super(string);
+    }
+}
