+++ src/main/java/tony/exceptions/ShowTasksException.java
@@ -1,0 +1,8 @@
+package tony.exceptions;
+
+public class ShowTasksException extends TonyException {
+
+    public ShowTasksException(String string) {
+        super(string);
+    }
+}
